+++ packages/events/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @happyvertical/smrt-events
 
+## 0.29.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.29.29
+  - @happyvertical/smrt-assets@0.29.29
+  - @happyvertical/smrt-places@0.29.29
+  - @happyvertical/smrt-profiles@0.29.29
+  - @happyvertical/smrt-tenancy@0.29.29
+  - @happyvertical/smrt-types@0.29.29
+
 ## 0.29.28
 
 ### Patch Changes