+++ packages/languages/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @happyvertical/smrt-languages
 
+## 0.29.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.29.29
+  - @happyvertical/smrt-features@0.29.29
+  - @happyvertical/smrt-jobs@0.29.29
+  - @happyvertical/smrt-prompts@0.29.29
+  - @happyvertical/smrt-tenancy@0.29.29
+  - @happyvertical/smrt-config@0.29.29
+
 ## 0.29.28
 
 ### Patch Changes