+++ packages/social/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @happyvertical/smrt-social
 
+## 0.29.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.29.29
+  - @happyvertical/smrt-content@0.29.29
+  - @happyvertical/smrt-secrets@0.29.29
+  - @happyvertical/smrt-tenancy@0.29.29
+  - @happyvertical/smrt-video@0.29.29
+  - @happyvertical/smrt-svelte@0.29.29
+  - @happyvertical/smrt-config@0.29.29
+
 ## 0.29.28
 
 ### Patch Changes